Fix subtract underflow in softpwm
Fixes #119

+++ src/gpio/soft_pwm.rs
@@ -12,13 +12,13 @@ use libc::{self, sched_param, timespec, CLOCK_MONOTONIC, PR_SET_TIMERSLACK, SCHE
 use super::{Error, GpioState, Result};
 
 // Only call sleep() if we have enough time remaining
-const SLEEP_THRESHOLD: u64 = 250_000;
+const SLEEP_THRESHOLD: i64 = 250_000;
 // Reserve some time for busy waiting
-const BUSYWAIT_MAX: u64 = 200_000;
+const BUSYWAIT_MAX: i64 = 200_000;
 // Subtract from the remaining busy wait time to account for get_time_ns() overhead
-const BUSYWAIT_REMAINDER: u64 = 100;
+const BUSYWAIT_REMAINDER: i64 = 100;
 
-const NANOS_PER_SEC: u64 = 1_000_000_000;
+const NANOS_PER_SEC: i64 = 1_000_000_000;
 
 #[derive(Debug, PartialEq, Eq, Copy, Clone)]
 enum Msg {
@@ -74,8 +74,8 @@ impl SoftPwm {
                 libc::prctl(PR_SET_TIMERSLACK, 1);
             }
 
-            let mut period_ns = period.as_nanos() as u64;
-            let mut pulse_width_ns = pulse_width.as_nanos() as u64;
+            let mut period_ns = period.as_nanos() as i64;
+            let mut pulse_width_ns = pulse_width.as_nanos() as i64;
 
             let mut start_ns = get_time_ns();
 
@@ -88,7 +88,7 @@ impl SoftPwm {
                 // Sleep if we have enough time remaining, while reserving some time
                 // for busy waiting to compensate for sleep taking longer than needed.
                 if pulse_width_ns >= SLEEP_THRESHOLD {
-                    sleep(Duration::from_nanos(pulse_width_ns - BUSYWAIT_MAX));
+                    sleep(Duration::from_nanos((pulse_width_ns - BUSYWAIT_MAX) as u64));
                 }
 
                 // Busy-wait for the remaining active time, minus BUSYWAIT_REMAINDER
@@ -106,8 +106,8 @@ impl SoftPwm {
                     match msg {
                         Msg::Reconfigure(period, pulse_width) => {
                             // Reconfigure period and pulse width
-                            pulse_width_ns = pulse_width.as_nanos() as u64;
-                            period_ns = period.as_nanos() as u64;
+                            pulse_width_ns = pulse_width.as_nanos() as i64;
+                            period_ns = period.as_nanos() as i64;
 
                             if pulse_width_ns > period_ns {
                                 pulse_width_ns = period_ns;
@@ -125,7 +125,7 @@ impl SoftPwm {
                 // Sleep if we have enough time remaining, while reserving some time
                 // for busy waiting to compensate for sleep taking longer than needed.
                 if remaining_ns >= SLEEP_THRESHOLD {
-                    sleep(Duration::from_nanos(remaining_ns - BUSYWAIT_MAX));
+                    sleep(Duration::from_nanos((remaining_ns - BUSYWAIT_MAX) as u64));
                 }
 
                 // Busy-wait for the remaining inactive time, minus BUSYWAIT_REMAINDER
@@ -179,7 +179,7 @@ impl Drop for SoftPwm {
 unsafe impl Sync for SoftPwm {}
 
 #[inline(always)]
-fn get_time_ns() -> u64 {
+fn get_time_ns() -> i64 {
     let mut ts = timespec {
         tv_sec: 0,
         tv_nsec: 0,
@@ -189,5 +189,5 @@ fn get_time_ns() -> u64 {
         libc::clock_gettime(CLOCK_MONOTONIC, &mut ts);
     }
 
-    (ts.tv_sec as u64 * NANOS_PER_SEC) + ts.tv_nsec as u64
+    (ts.tv_sec as i64 * NANOS_PER_SEC) + ts.tv_nsec as i64
 }
